+++ src/thttprequest.cpp
@@ -1,4 +1,4 @@
-/* Copyright (c) 2010-2015, AOYAMA Kazuharu
+/* Copyright (c) 2010-2017, AOYAMA Kazuharu
  * All rights reserved.
  *
  * This software may be used and distributed according to the terms of
@@ -342,7 +342,9 @@ QVariantMap THttpRequest::formItems(const QString &key) const
 void THttpRequest::parseBody(const QByteArray &body, const THttpRequestHeader &header)
 {
     switch (method()) {
-    case Tf::Post: {
+    case Tf::Post:
+    case Tf::Put:
+    case Tf::Patch: {
         QString ctype = QString::fromLatin1(header.contentType().trimmed());
         if (ctype.startsWith("multipart/form-data", Qt::CaseInsensitive)) {
             // multipart/form-data
@@ -366,8 +368,7 @@ void THttpRequest::parseBody(const QByteArray &body, const THttpRequestHeader &h
 #else
             tSystemWarn("unsupported content-type: %s", qPrintable(ctype));
 #endif
-        } else {
-            // 'application/x-www-form-urlencoded'
+        } else if (ctype.startsWith("application/x-www-form-urlencoded", Qt::CaseInsensitive)) {
             if (!body.isEmpty()) {
                 const QList<QByteArray> formdata = body.split('&');
                 for (auto &frm : formdata) {
@@ -377,10 +378,12 @@ void THttpRequest::parseBody(const QByteArray &body, const THttpRequestHeader &h
                         QString key = THttpUtility::fromUrlEncoding(nameval.value(0));
                         QString val = THttpUtility::fromUrlEncoding(nameval.value(1));
                         d->formItems.insertMulti(key, val);
-                        tSystemDebug("POST Hash << %s : %s", qPrintable(key), qPrintable(val));
+                        tSystemDebug("x-www-form-urlencoded << %s : %s", qPrintable(key), qPrintable(val));
                     }
                 }
             }
+        } else {
+            tSystemWarn("unsupported content-type: %s", qPrintable(ctype));
         }
         /* FALL THROUGH */ }
 
@@ -421,8 +424,12 @@ QByteArray THttpRequest::boundary() const
         for (auto &bnd : lst) {
             QString string = bnd.trimmed();
             if (string.startsWith("boundary=", Qt::CaseInsensitive)) {
-                boundary  = "--";
-                boundary += string.mid(9).toLatin1();
+                boundary = string.mid(9).toLatin1();
+                // strip optional surrounding quotes (RFC 2046 and 7578)
+                if (boundary.startsWith('"') && boundary.endsWith('"')) {
+                    boundary = boundary.mid(1, boundary.size() - 2);
+                }
+                boundary.prepend("--");
                 break;
             }
         }