+++ src/tjsloader.cpp
@@ -1,4 +1,4 @@
-/* Copyright (c) 2016, AOYAMA Kazuharu
+/* Copyright (c) 2016-2017, AOYAMA Kazuharu
  * All rights reserved.
  *
  * This software may be used and distributed according to the terms of
@@ -239,7 +239,7 @@ void TJSLoader::import(const QString &defaultMember, const QString &moduleName)
             return;
         }
     }
-    importFiles << QPair<QString, QString>(defaultMember, moduleName);
+    importFiles << qMakePair(defaultMember, moduleName);
 }