+++ tools/tspawn/validatorgenerator.cpp
@@ -1,4 +1,4 @@
-/* Copyright (c) 2011-2015, AOYAMA Kazuharu
+/* Copyright (c) 2011-2017, AOYAMA Kazuharu
  * All rights reserved.
  *
  * This software may be used and distributed according to the terms of
@@ -22,6 +22,8 @@
     "public:\n"                                                         \
     "    %2Validator();\n"                                              \
     "};\n"                                                              \
+    "\n"                                                                \
+    "Q_DECLARE_METATYPE(%2Validator)\n"                                 \
     "\n"                                                                \
     "#endif // %1VALIDATOR_H\n"